- dumps for radau/lobatto solvers
git-svn-id: https://openmodelica.org/svn/OpenModelica/trunk@16263 f25d12d1-65f4-0310-ae8a-bbce733d8d8e

+++ SimulationRuntime/c/simulation/solver/radau.h
@@ -78,7 +78,7 @@
       double *max;
       double *derx;
       double *s;
-      double **c;
+      long double **c;
       double *a;
     }NLPODE;
